requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
fastapi==0.112.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
inflection==0.5.1
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mypy-extensions==1.0.0
openai==1.43.0
openapi==2.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.9
referencing==0.35.1
rpds-py==0.20.0
sniffio==1.3.1
starlette==0.38.4
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.30.6
